% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsearchR.R
\name{dsearchR}
\alias{dsearchR}
\title{Find pairs of points within fixed cutoff}
\usage{
dsearchR(pts, r)
}
\arguments{
\item{pts}{A list of coordinate vectors (of any dimension).}

\item{r}{The cutoff distance criteria between points (<=r).}
}
\value{
A list of length 2 vectors.
}
\description{
Find pairs of points within fixed cutoff
}
\examples{
pts <- list(c(1,1), c(2,1), c(2,2), c(1,2))
pairs <- dsearchR(pts, 1.1)
}
