% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dsearchRcpp}
\alias{dsearchRcpp}
\title{Find pairs of points within fixed cutoff}
\usage{
dsearchRcpp(dim, pts, r)
}
\arguments{
\item{dim}{The number of spatial dimensions}

\item{pts}{A list of coordinate vectors (of length dim).}

\item{r}{The cutoff distance criteria between points (<=r).}
}
\value{
A list of length 2 vectors.
}
\description{
Find pairs of points within fixed cutoff
}
\examples{
pts <- list(c(1,1), c(2,1), c(2,2), c(1,2))
pairs <- dsearchRcpp(2, pts, 1.1)
}
